+++ primint/src/num/nonmax.rs
@@ -1,3 +1,4 @@
+use core::fmt::{Display, Formatter};
 use crate::UnsignedPrimInt;
 use crate::num::NonZero;
 
@@ -86,6 +87,11 @@ impl<T: UnsignedPrimInt> Default for NonMax<T> {
     fn default() -> Self {
         // SAFETY: We know that zero is not the maximum value for any primint
         unsafe { Self::new_unchecked(T::ONE) }
+    }
+}
+impl<T: UnsignedPrimInt> Display for NonMax<T> {
+    fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
+        write!(f, "{}", self.get())
     }
 }